# Detailed/03-OpenAPI/med-rms-openapi.yaml
...
openapi: 3.0.3
info:
  title: Med-RMS 医疗器械需求管理系统 API
  description: 'Med-RMS (Medical Device Requirement Management System) RESTful API 规范。

    涵盖需求管理、追溯管理、变更管理、合规管理、电子签名、风险管理、项目管理、报表仪表盘及系统管理模块。

    '
  version: 1.1.0
  contact:
    name: Med-RMS 开发团队
  license:
    name: Proprietary
servers:
- url: http://localhost:8080/api/v1
  description: 开发环境
...
security:
- BearerAuth: []
paths:
  /auth/login:
    post:
      operationId: login
      tags:
      - 认证与用户
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDTO'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/TokenVO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证（Token缺失或过期）
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: 系统内部错误
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: success
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
...
    ErrorResponse:
      type: object
      required:
      - errorCode
      - errorMessage
      properties:
        errorCode:
          type: string
          description: 错误码
          example: SY0100
        errorMessage:
          type: string
          description: 错误消息
          example: 请求参数校验失败
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: 字段级错误详情
...
    LoginDTO:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          description: 用户名
          maxLength: 50
        password:
          type: string
          description: 密码
          format: password
          maxLength: 128
...
    TokenVO:
      type: object
      required:
      - accessToken
      - refreshToken
      - expiresIn
      - tokenType
      properties:
        accessToken:
          type: string
          description: 访问令牌
        refreshToken:
          type: string
          description: 刷新令牌
        expiresIn:
          type: integer
          description: 过期时间（秒）
          example: 900
        tokenType:
          type: string
          description: 令牌类型
          example: Bearer
...